# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para el sistema de monitoreo de instancias Proxmox El Centinela.\n**Flujo de autenticación**: Login → `/auth/login` → obtener `jwtTemporal` → `/auth/2fa/verify` → obtener `accessToken` → usar en el botón **Authorize** de esta UI.",
        "title": "El Centinela API",
        "contact": {
            "name": "Soporte El Centinela"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/events": {
            "get": {
                "description": "Abre un stream text/event-stream. Requiere el ticket de POST /api/events/ticket, que se consume al conectar (un solo uso). Cada evento llega como `data: \u003cRealtimeEvent en JSON\u003e`. Si el backend corta el stream (logout, revocación de sesiones o usuario desactivado) envía antes `event: cierre` con `{\"motivo\": \"...\"}`. Cada 25 s llega un comentario `: ping` para mantener la conexión.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Eventos en tiempo real"
                ],
                "summary": "Stream de eventos en tiempo real (SSE)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket de un solo uso",
                        "name": "ticket",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stream de eventos",
                        "schema": {
                            "$ref": "#/definitions/ports.RealtimeEvent"
                        }
                    },
                    "401": {
                        "description": "EVENTS_TICKET_MISSING | EVENTS_TICKET_INVALID",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/ticket": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve un ticket de un solo uso, válido por 30 segundos, para abrir GET /api/events?ticket=... El navegador (EventSource) no puede mandar el header Authorization, y así el JWT nunca viaja en la URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos en tiempo real"
                ],
                "summary": "Pedir ticket para el stream de eventos",
                "responses": {
                    "200": {
                        "description": "ticket",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Mensaje descriptivo del error."
                }
            }
        },
...
        "ports.RealtimeEvent": {
            "description": "Esquema genérico de evento del canal en tiempo real (RF-11). Los enums usan valores en inglés MAYÚSCULA y los campos nombres en español camelCase.",
            "type": "object",
            "properties": {
                "detalles": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "fechaHora": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mensaje": {
                    "type": "string"
                },
                "recursoId": {
                    "type": "string"
                },
                "recursoTipo": {
                    "type": "string"
                },
                "severidad": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de acceso definitivo. Formato: **Bearer \u0026lt;accessToken\u0026gt;**. Obtené el token completando el flujo: `POST /auth/login` → `POST /auth/2fa/verify`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerPreAuth": {
            "description": "JWT temporal pre-2FA. Formato: **Bearer \u0026lt;jwtTemporal\u0026gt;**. Obtené el token de `POST /auth/login`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}